FIX: IsCompany int64 to bool

+++ db/schema/schema.sql
@@ -818,7 +818,7 @@ CREATE TABLE Transactant (
     LastName VARCHAR(100) NOT NULL DEFAULT '',
     PreferredName VARCHAR(100) NOT NULL DEFAULT '',
     CompanyName VARCHAR(100) NOT NULL DEFAULT '',
-    IsCompany SMALLINT NOT NULL DEFAULT 0,                  -- 0 == this is a person,  1 == this is a company
+    IsCompany TINYINT(1) NOT NULL DEFAULT 0,                  -- 0 == this is a person,  1 == this is a company
     PrimaryEmail VARCHAR(100) NOT NULL DEFAULT '',
     SecondaryEmail VARCHAR(100) NOT NULL DEFAULT '',
     WorkPhone VARCHAR(100) NOT NULL DEFAULT '',
@@ -1407,7 +1407,7 @@ CREATE TABLE ClosePeriod (
 CREATE TABLE Flow (
     FlowID BIGINT NOT NULL AUTO_INCREMENT,
     BID BIGINT NOT NULL DEFAULT 0,                                                         -- Business id
-    UserRefNo VARCHAR(50) NOT NULL DEFAULT '',                                             -- reference id to share with the user(s)    
+    UserRefNo VARCHAR(50) NOT NULL DEFAULT '',                                             -- reference id to share with the user(s)
     FlowType VARCHAR(50) NOT NULL DEFAULT '',                                              -- for which flow we're storing data ("RA=Rental Agreement Flow")
     Data JSON DEFAULT NULL,                                                                -- JSON Data for each flow type
     LastModTime TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,  -- when was it last updated
